requirements.txt
certifi
chardet
docutils
idna
Kivy
kivy-deps.glew
kivy-deps.gstreamer
kivy-deps.sdl2
Kivy-Garden
kivymd
numpy
opencv-python
pickle-mixin
Pillow
Pygments
pypiwin32
pywin32
requests
six
SQLAlchemy
SQLAlchemy-Utils
urllib3
python-dotenv